Optimize: 	1. Reorganize the project by new Makefile flavor. 	2. Adding the local vscode settings.
