--- inc/client/mqttor_client.h
@@ -1,20 +1,0 @@
-/*! \brief a demo matt client
- *  \author Shylock Hg
- *  \date 2018-08-13
- *  \email tcath2s@gmail.com
- * */
-
-#ifndef _MQTTOR_CLIENT_H_
-#define _MQTTOR_CLIENT_H_
-
-#ifdef __cplusplus
-	extern "C" {
-#endif
-
-
-#ifdef __cplusplus
-	}
-#endif
-
-#endif  //!< _MQTTOR_CLIENT_H_
-
