+++ mqtt/inc/mqtt_packet_segment.h
