+++ mqtt/inc/mqttor_session.h
@@ -13,7 +13,7 @@
 	
 #include <stdbool.h>
 
-#include <core/mqtt_packet.h>
+#include "./mqtt_packet.h"
 
 enum mqttor_protocol {
 	MQTTOR_V311 = 4
