+++ mqtt/src/mqtt_packet_segment.c
@@ -10,10 +10,10 @@
 #include <assert.h>
 #include <stdio.h>
 
-#include <core/mqtt_packet_segment.h>
+#include "../inc/mqtt_packet_segment.h"
 
-#include <toolkit/array.h>
-#include <toolkit/mqtt_log.h>
+#include "../inc/array.h"
+#include "../inc/mqtt_log.h"
 
 
 //
