+++ mqtt/src/mqtt_var_header.c
@@ -1,6 +1,6 @@
 #include <stdint.h>
 
-#include <core/mqtt_var_header.h>
+#include "../inc/mqtt_var_header.h"
 
 struct mqtt_buf_connect_flag * mqtt_connect_flag_pack(
 		union mqtt_attr_connect_flag flag){
