+++ mqtt/test.c
@@ -13,13 +13,11 @@
 #include <arpa/inet.h>
 #include <unistd.h>
 
-#include "core/mqtt_fixed_header.h"
-#include "core/mqtt_var_header.h"
-#include "core/mqtt_payload.h"
-
-#include "core/mqtt_packet.h"
-
-#include "toolkit/mqtt_log.h"
+#include "./inc/mqtt_fixed_header.h"
+#include "./inc/mqtt_var_header.h"
+#include "./inc/mqtt_payload.h"
+#include "./inc/mqtt_packet.h"
+#include "./inc/mqtt_log.h"
 
 void test_toolkit(void);
 void test_packet_segment(void);
